Rename node encoder files
