+++ packages/dds/tree/src/feature-libraries/chunked-forest/codec/nodeEncoder.ts
